LayoutTests/platform/ios-simulator/fast/events/clientXY-in-zoom-and-scroll-expected.txt

Base
PASS event.clientX is 100
PASS event.clientY is 100
PASS event.pageX is 100
PASS event.pageY is 100

Just zoomed
PASS event.clientX is 83
PASS event.clientY is 83
PASS event.pageX is 83
PASS event.pageY is 83

Just scrolled
PASS event.clientX is 100
PASS event.clientY is 100
FAIL event.pageX should be 150. Was 100.
FAIL event.pageY should be 150. Was 100.

Zoomed and scrolled
PASS event.clientX is 83
PASS event.clientY is 83
FAIL event.pageX should be 133. Was 83.
FAIL event.pageY should be 133. Was 83.
PASS successfullyParsed is true

TEST COMPLETE
